# backend/orders/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package http

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Address defines model for Address.
type Address struct {
	// City City of the address
	City string `json:"city"`

	// CountryCode Country code in ISO 3166-1 alpha-2 format
	CountryCode CountryCode `json:"country_code"`

	// Line1 First line of the address
	Line1 string `json:"line1"`

	// Line2 Second line of the address
	Line2 string `json:"line2"`

	// PostalCode Postal code of the address
	PostalCode string `json:"postal_code"`
}

// CountryCode Country code in ISO 3166-1 alpha-2 format
type CountryCode = string

// CustomerUUID UUID of a customer
type CustomerUUID = openapi_types.UUID

// ErrorDetails defines model for ErrorDetails.
type ErrorDetails struct {
	ErrorId   *string `json:"error_id,omitempty"`
	ErrorSlug string  `json:"error_slug"`
	ErrorType *string `json:"error_type,omitempty"`
	Message   string  `json:"message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Details Detailed error information
	Details []ErrorDetails `json:"details"`

	// Message Human-readable error message
	Message string `json:"message"`

	// Slug Error slug
	Slug string `json:"slug"`
}

// RegisterCustomer defines model for RegisterCustomer.
type RegisterCustomer struct {
	Address Address `json:"address"`

	// Email Customer email address
	Email openapi_types.Email `json:"email"`

	// Name Customer's full name
	Name string `json:"name"`

	// PhoneNumber Customer phone number
	PhoneNumber string `json:"phone_number"`
}

// RegisterCustomerResponse defines model for RegisterCustomerResponse.
type RegisterCustomerResponse struct {
	// CustomerUuid UUID of a customer
	CustomerUuid CustomerUUID `json:"customer_uuid"`
}

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// RegisterCustomerJSONRequestBody defines body for RegisterCustomer for application/json ContentType.
type RegisterCustomerJSONRequestBody = RegisterCustomer

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Register a new customer
	// (POST /orders/register-customer)
	RegisterCustomer(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// RegisterCustomer converts echo context to params.
func (w *ServerInterfaceWrapper) RegisterCustomer(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RegisterCustomer(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(options.BaseURL+"/orders/register-customer", wrapper.RegisterCustomer, options.OperationMiddlewares["registerCustomer"]...)

}

type BadRequestJSONResponse ErrorResponse

type RegisterCustomerRequestObject struct {
	Body *RegisterCustomerJSONRequestBody
}

type RegisterCustomerResponseObject interface {
	VisitRegisterCustomerResponse(w http.ResponseWriter) error
}

type RegisterCustomer201JSONResponse RegisterCustomerResponse

func (response RegisterCustomer201JSONResponse) VisitRegisterCustomerResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type RegisterCustomer400JSONResponse struct{ BadRequestJSONResponse }

func (response RegisterCustomer400JSONResponse) VisitRegisterCustomerResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type RegisterCustomer409JSONResponse ErrorResponse

func (response RegisterCustomer409JSONResponse) VisitRegisterCustomerResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Register a new customer
	// (POST /orders/register-customer)
	RegisterCustomer(ctx context.Context, request RegisterCustomerRequestObject) (RegisterCustomerResponseObject, error)
}

type StrictHandlerFunc func(ctx echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// RegisterCustomer operation middleware
func (sh *strictHandler) RegisterCustomer(ctx echo.Context) error {
	var request RegisterCustomerRequestObject

	var body RegisterCustomerJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RegisterCustomer(ctx.Request().Context(), request.(RegisterCustomerRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RegisterCustomer")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(RegisterCustomerResponseObject); ok {
		return validResponse.VisitRegisterCustomerResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}